Moved from app.rs to routes.rs for router
