--- src/infrastructure/mod.rs
@@ -1,3 +1,0 @@
-pub mod logging;
-pub mod repository;
-pub mod server;
