+++ src/lib.rs
@@ -1,4 +1,4 @@
-mod app;
+mod routes;
 
 mod functions;
 mod layouts;
@@ -21,7 +21,7 @@ if #[cfg(feature = "hydrate")] {
 
     #[wasm_bindgen]
     pub fn hydrate() {
-      use app::*;
+      use routes::*;
 
       console_error_panic_hook::set_once();
